Merged data preparation
I will re-run GridSearchCV and RandomizedSearchCV to find the best parameters for the first four models as in old Notebook . Then, I will try to run NN and create a prediction for test dataset. Not sure if I can finish today because my Jupyter sometimes raises an exception and is closed. I'll do my best.